Initial creation of media
The wall model, texture, and several normal maps were created. Some of
the normal maps have media that was used to create them.

The normal maps are:

1. Generated from edge detection with the GIMP's difference of gaussians
   filter and then converted to normal map. This normal map seems to
   have been inverted by accident.
2. Generated by the GIMP's Normalmap filter w/ default settings.
3. Hand generated by tracing the lines between bricks. See the heightmap
   in the sourceFiles folder.
4. The height map from #3 was given a gaussian blur.

The edge detected image from #1 and the height maps from #3 and 4 were
passed to http://cpetry.github.io/NormalMap-Online/ to be converted to
a normal map (default settings used).

Note that the influence given to the normal map in Blender varies. Some
maps are very sensitive and must be given lower weights (eg, I rendered
the animation for #2 with 0.5 weight but #3 was given 5 weight).